fix(cli): rename `setup --domain` argument to `--api-domain`
BREAKING CHANGE

Turns out older versions of Node.js core didn't like our previous
`--domain` command line argument. In short it makes Node.js core think
it's dealing with a legacy Node.js Domain instance, and therefore ends
up trying to invoke `Domain.prototype.enter` which doesn't exist.

This is no longer an issue with more recent versions of
Node.js (10 and above), but since we've obviously got users of
`gitlab-search` tripping into this weird bug with older versions of
Node.js installed, we might as well do them a favor and avoid it from
the get-go.

Therefore renaming the old `--domain` argument to `--api-domain`.

Worth mentioning that we still persist the given value as `domain` in
`.gitlabsearchrc` primarily to avoid causing havoc for users who have
already configured `gitlab-search` correctly.

The fact that the argument is called `"api-domain"` but persisted as
`"domain"` can be _slightly_ confusing, but personally don't think it
matters much as long as users are able to setup it up from the command
line, rather than diving in and craft the `.gitlabsearchrc` file by hand.

Closes https://github.com/phillipj/gitlab-search/issues/9