chore: use ?per_page=100 in requests since 100 is GitLab's limit
As we found when debugging why certain projects wasn't included when
searching in certain scenarios, we found that GitLab has a max limit of
`100` when fetching groups, projects and the like.

With that in mind, it is very confusing that gitlab-search tried to
use `?per_page=1000`, knowing it would never be possible to retrieve
more than `100` per request.

Refs https://github.com/phillipj/gitlab-search/issues/14